backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.2.10
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
Django==3.0.8
django-cors-headers==3.4.0
django-rest-framework==0.1.0
djangorestframework==3.11.2
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
openai==0.27.8
pytz==2020.1
requests==2.31.0
sqlparse==0.3.1
tqdm==4.66.1
urllib3==2.0.4
yarl==1.9.2